requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-shell-plus==1.1.7